+++ core/server/data/export/index.js
@@ -2,15 +2,13 @@ var _          = require('lodash'),
     when       = require('when'),
     versioning = require('../versioning'),
     knex       = require('../../models/base').knex,
-    schema     = require('../schema').tables,
+    utils      = require('../utils'),
 
     excludedTables = ['sessions'],
     exporter;
 
 exporter = function () {
-    var tablesToExport = _.keys(schema);
-
-    return when.join(versioning.getDatabaseVersion(), tablesToExport).then(function (results) {
+    return when.join(versioning.getDatabaseVersion(), utils.getTables()).then(function (results) {
         var version = results[0],
             tables = results[1],
             selectOps = _.map(tables, function (name) {
@@ -36,7 +34,7 @@ exporter = function () {
 
             return when.resolve(exportData);
         }, function (err) {
-            console.log("Error exporting data: " + err);
+            console.log('Error exporting data: ' + err);
         });
     });
 };
